change data type to field status in booking table

+++ migrations/20190823191024-create-booking.js
@@ -34,7 +34,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       status: {
-        type: Sequelize.TINYINT
+        type: Sequelize.INTEGER(2)
       },
       createdAt: {
         allowNull: false,
